Backend/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
h11==0.13.0
idna==3.3
oauthlib==3.2.0
outcome==1.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.10
wsproto==1.1.0
